blocks-move-test-data - move test_data under gradio
