+++ gradio/test_data/lion.jpg
