+++ gradio/test_data/test_image.png
