+++ test/test_blocks.py
@@ -1,8 +1,8 @@
 import random
 import unittest
-from test.test_data.blocks_configs import XRAY_CONFIG
 
 import gradio as gr
+from gradio.test_data.blocks_configs import XRAY_CONFIG
 
 
 class TestBlocks(unittest.TestCase):
