+++ test/test_encryptor.py
@@ -1,8 +1,8 @@
 import os
 import unittest
-from test.test_data.media_data import BASE64_IMAGE
 
 from gradio import encryptor, processing_utils
+from gradio.test_data.media_data import BASE64_IMAGE
 
 os.environ["GRADIO_ANALYTICS_ENABLED"] = "False"
